backend/requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.72
botocore==1.31.72
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
debugpy==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.104.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
pyotp==2.9.0
pytest==7.4.3
httpx==0.25.1